requirements.txt
streamlit>=1.31.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.15
openai>=1.12.0
langchain-openai
langchain-community
langdetect
faiss-cpu
langchain-pinecone
pypdf
tiktoken